requirements.txt
matplotlib==3.8.2
numpy==1.24.3
pydantic==2.7.1
PyYAML==6.0.1
tensorflow==2.13.0
typing_extensions==4.11.0
